Now dumps LevelDatabase, AssetDatabase, Wwise stuff
Serialization in LevelData is now an important thing to discern, it's a binary format stream that is populated on runtime. It would help a TON to learn what was in it, which seems to be a list of TrackSections that gets put into the sections array of LevelData (perhaps it populates the fields of the already existing objects? Maybe it adds more? Sets?)